WIP: Removing all vestiges of vector_engine
